wip: Create addin api crate
