+++ governance/addin-api/src/lib.rs
@@ -1,0 +1,3 @@
+//! Governance add-ins interfaces
+pub mod max_voter_weight;
+pub mod voter_weight;
