+++ governance/program/Cargo.toml
@@ -22,6 +22,7 @@ serde_derive = "1.0.103"
 solana-program = "1.9.5"
 spl-token = { version = "3.3", path = "../../token/program", features = [ "no-entrypoint" ] }
 spl-governance-tools= { version = "0.1.0", path ="../tools"}
+spl-governance-addin-api= { version = "0.1.0", path ="../addin-api"}
 thiserror = "1.0"
 
 [dev-dependencies]
