+++ governance/program/tests/use_realm_with_voter_weight_addin.rs
@@ -7,10 +7,10 @@ mod program_test;
 
 use program_test::*;
 use spl_governance::{
-    addins::voter_weight::VoterWeightAction,
     error::GovernanceError,
     state::vote_record::{Vote, VoteChoice},
 };
+use spl_governance_addin_api::voter_weight::VoterWeightAction;
 
 #[tokio::test]
 async fn test_create_account_governance_with_voter_weight_addin() {
